refactor(core/event): move ontology under services/event/ontologies/ per convention
onto2py emits two outputs: a consolidated module file and a per-class
directory tree under ontologies/classes/. The per-class tree was missing
on first pass because I'd been running onto2py with the TTL at
services/event/event.ttl, which doesn't contain an "ontologies" path
segment — so onto2py fell back to libs/.../services/ontologies/classes/,
which I deleted as an apparent orphan.

Aligning with the convention used by BFO7BucketsProcessOntology in
libs/naas-abi/naas_abi/ontologies/:

  services/event/ontologies/modules/EventOntology.ttl
  services/event/ontologies/modules/EventOntology.py    (consolidated)
  services/event/ontologies/classes/ontology_naas_ai/abi/LogProcess.py
  services/event/ontologies/classes/purl_obolibrary_org/obo/Process.py

Imports updated in EventService, tests, benchmark, and EventPort.md.
37 tests pass; make check-core green.
