+++ libs/naas-abi-core/naas_abi_core/services/event/ontologies/classes/ontology_naas_ai/abi/LogProcess.py
@@ -1,0 +1,11 @@
+from naas_abi_core.services.event.ontologies.modules.EventOntology import (
+    LogProcess as _LogProcess,
+)
+
+
+class LogProcess(_LogProcess):
+    """Action class for LogProcess"""
+
+    def actions(self):
+        """Action method - implement your logic here"""
+        pass
