+++ libs/naas-abi-core/naas_abi_core/services/event/ontologies/classes/purl_obolibrary_org/obo/Process.py
@@ -1,0 +1,11 @@
+from naas_abi_core.services.event.ontologies.modules.EventOntology import (
+    Process as _Process,
+)
+
+
+class Process(_Process):
+    """Action class for Process"""
+
+    def actions(self):
+        """Action method - implement your logic here"""
+        pass
