+++ libs/naas-abi-core/naas_abi_core/services/event/ontologies/modules/EventOntology.ttl
